# schemas/api.v1.openapi.json
{
  "openapi": "3.1.0",
  "jsonSchemaDialect": "https://json-schema.org/draft/2020-12/schema",
  "info": {
    "title": "AgentMash Backend API",
    "version": "1.0.0",
    "description": "Implemented API contract for submitting agent artifacts, pulling a human review queue, publishing reviewed feedback bundles, reading feedback by run ID, and deleting artifacts."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin AgentMash API server"
    }
  ],
  "x-agentmash-status": "implemented",
  "x-agentmash-runtime": "server/agentmash-api.mjs",
  "x-agentmash-user-action": "Set AGENTMASH_API_TOKEN before production use and serve the PWA through the API server or a same-origin reverse proxy.",
  "x-agentmash-example-files": [
    "/schemas/examples/intake.v1.json",
    "/schemas/examples/intake-ack.v1.json",
    "/schemas/examples/feedback-bundle.v1.json"
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/v1/health": {
      "get": {
        "operationId": "getApiHealth",
        "summary": "Check API health and intake limits",
        "security": [],
        "responses": {
          "200": {
            "description": "API is ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/v1/intake": {
      "post": {
        "operationId": "submitArtifactsForReview",
        "summary": "Submit artifacts for human first-impression review",
        "description": "Accepts the same agentmash.intake.v1 payload that the app imports as a local agent drop, validates it, stores image data separately, and queues accepted artifacts for reviewers.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./intake.v1.json"
              },
              "examples": {
                "localAgentDrop": {
                  "summary": "Local agent-drop payload",
                  "externalValue": "/schemas/examples/intake.v1.json"
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Artifacts accepted for review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeAcknowledgement"
                },
                "examples": {
                  "accepted": {
                    "summary": "Accepted intake acknowledgement",
                    "externalValue": "/schemas/examples/intake-ack.v1.json"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Payload failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication"
          },
          "413": {
            "description": "Payload or image data exceeded configured limits"
          },
          "429": {
            "description": "Submission rate limit exceeded"
          }
        }
      }
    },
    "/v1/review-queue": {
      "get": {
        "operationId": "getReviewQueue",
        "summary": "Pull artifacts into the human review deck",
        "description": "Returns queued artifacts, optionally with image data URLs, so the PWA can insert submitted artifacts into the human swipe flow.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          },
          {
            "name": "includeImageData",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Queued artifacts for review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewQueue"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication"
          }
        }
      }
    },
    "/v1/feedback": {
      "post": {
        "operationId": "publishFeedbackBundle",
        "summary": "Publish reviewed feedback back to the API",
        "description": "Stores an agentmash.feedback-bundle.v1 bundle generated by the human review app so agents can fetch results by run ID.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackBundle"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Feedback bundle stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackAcknowledgement"
                }
              }
            }
          },
          "400": {
            "description": "Payload failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication"
          }
        }
      }
    },
    "/v1/feedback/{runId}": {
      "get": {
        "operationId": "getFeedbackBundle",
        "summary": "Read feedback packets for an agent run",
        "description": "Returns ready feedback packets, eval rows, and pairwise rows for a submitted run after the human review app publishes a feedback bundle.",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feedback rows are available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackBundle"
                },
                "examples": {
                  "readyBundle": {
                    "summary": "Ready feedback bundle",
                    "externalValue": "/schemas/examples/feedback-bundle.v1.json"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication"
          },
          "404": {
            "description": "Run was not found or is not visible to the caller"
          }
        }
      }
    },
    "/v1/artifacts/{artifactId}": {
      "delete": {
        "operationId": "deleteArtifact",
        "summary": "Delete a submitted artifact and related image bytes",
        "description": "Deletes a submitted artifact and related image bytes after bearer-token authentication.",
        "parameters": [
          {
            "name": "artifactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artifact deletion was accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionAcknowledgement"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication"
          },
          "404": {
            "description": "Artifact was not found or is not visible to the caller"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Health": {
        "type": "object",
        "required": ["schema", "status", "generatedAt", "limits"],
        "properties": {
          "schema": {
            "const": "agentmash.health.v1"
          },
          "status": {
            "const": "ok"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "limits": {
            "$ref": "#/components/schemas/IntakeLimits"
          }
        },
        "additionalProperties": false
      },
      "IntakeAcknowledgement": {
        "type": "object",
        "required": ["schema", "status", "accepted", "rejected", "limits"],
        "properties": {
          "schema": {
            "const": "agentmash.intake-ack.v1"
          },
          "status": {
            "enum": ["accepted", "partially_accepted", "rejected"]
          },
          "runId": {
            "type": "string"
          },
          "accepted": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AcceptedArtifact"
            }
          },
          "rejected": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            }
          },
          "limits": {
            "$ref": "#/components/schemas/IntakeLimits"
          }
        },
        "additionalProperties": false
      },
      "AcceptedArtifact": {
        "type": "object",
        "required": ["artifactId", "status"],
        "properties": {
          "artifactId": {
            "type": "string"
          },
          "status": {
            "enum": ["queued_for_review"]
          },
          "reviewContext": {
            "$ref": "./intake.v1.json#/$defs/reviewContext"
          }
        },
        "additionalProperties": false
      },
      "IntakeLimits": {
        "type": "object",
        "required": ["maxArtifactsPerRequest", "maxImageBytes", "allowedImageTypes"],
        "properties": {
          "maxArtifactsPerRequest": {
            "type": "integer",
            "minimum": 1
          },
          "maxImageBytes": {
            "type": "integer",
            "minimum": 1
          },
          "allowedImageTypes": {
            "type": "array",
            "items": {
              "enum": ["image/png", "image/jpeg", "image/webp"]
            }
          }
        },
        "additionalProperties": false
      },
      "ReviewQueue": {
        "type": "object",
        "required": ["schema", "generatedAt", "count", "artifacts", "limits"],
        "properties": {
          "schema": {
            "const": "agentmash.review-queue.v1"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "artifacts": {
            "type": "array",
            "items": {
              "$ref": "./intake.v1.json#/$defs/artifact"
            }
          },
          "limits": {
            "$ref": "#/components/schemas/IntakeLimits"
          }
        },
        "additionalProperties": false
      },
      "FeedbackBundle": {
        "type": "object",
        "required": ["schema", "runId", "status", "summary", "packets", "evalRows", "pairwiseRows"],
        "properties": {
          "schema": {
            "const": "agentmash.feedback-bundle.v1"
          },
          "runId": {
            "type": "string"
          },
          "status": {
            "enum": ["empty", "pending", "partial", "ready", "not_found"]
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "summary": {
            "$ref": "#/components/schemas/FeedbackBundleSummary"
          },
          "packets": {
            "type": "array",
            "items": {
              "$ref": "./feedback.v2.json"
            }
          },
          "evalRows": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "pairwiseRows": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "FeedbackBundleSummary": {
        "type": "object",
        "required": [
          "artifactCount",
          "reviewedArtifacts",
          "pendingArtifacts",
          "packetCount",
          "evalRowCount",
          "pairwiseRowCount",
          "runIds",
          "reviewers"
        ],
        "properties": {
          "artifactCount": {
            "type": "integer",
            "minimum": 0
          },
          "reviewedArtifacts": {
            "type": "integer",
            "minimum": 0
          },
          "pendingArtifacts": {
            "type": "integer",
            "minimum": 0
          },
          "packetCount": {
            "type": "integer",
            "minimum": 0
          },
          "evalRowCount": {
            "type": "integer",
            "minimum": 0
          },
          "pairwiseRowCount": {
            "type": "integer",
            "minimum": 0
          },
          "runIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reviewers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "FeedbackAcknowledgement": {
        "type": "object",
        "required": ["schema", "status", "storedAt", "runIds", "packetCount", "evalRowCount", "pairwiseRowCount"],
        "properties": {
          "schema": {
            "const": "agentmash.feedback-ack.v1"
          },
          "status": {
            "const": "stored"
          },
          "storedAt": {
            "type": "string",
            "format": "date-time"
          },
          "runIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "packetCount": {
            "type": "integer",
            "minimum": 0
          },
          "evalRowCount": {
            "type": "integer",
            "minimum": 0
          },
          "pairwiseRowCount": {
            "type": "integer",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "DeletionAcknowledgement": {
        "type": "object",
        "required": ["schema", "artifactId", "status"],
        "properties": {
          "schema": {
            "const": "agentmash.deletion-ack.v1"
          },
          "artifactId": {
            "type": "string"
          },
          "status": {
            "enum": ["deleted", "already_absent"]
          }
        },
        "additionalProperties": false
      },
      "ValidationError": {
        "type": "object",
        "required": ["schema", "message", "issues"],
        "properties": {
          "schema": {
            "const": "agentmash.validation-error.v1"
          },
          "message": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            }
          }
        },
        "additionalProperties": false
      },
      "ValidationIssue": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "path": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}